Finishing touches.
 - Move around few more functions from and to util.rs.
 - More explainer and todo docs.
 - Fixing cargo clippy for taker_protocol.rs.
